pi/pyproject.toml
[project]
name = "polyumi_pi"
version = "0.1.0"
description = "Raspberry Pi streaming client for the PolyUMI multimodal learning platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.3",
    "protobuf>=7.34.0",
    "pyzmq>=27.1.0",
    "rpi-hardware-pwm>=0.3.1",
    "sounddevice>=0.5.5",
    "typer>=0.24.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
# this should be installed using apt on the Pi, not pip
# but for development convenience we can install it in the 
# PC virtualenv here with uv
dev = [
    "picamera2",
    "pytest>=9.0.2",
    ]
